프로젝트 내에서 마일스톤, 태그를 조회하는 기능을 구현했습니다.

+++ task/src/main/java/com/kimtaehoonki/task/milestone/application/MilestoneServiceImpl.java
@@ -1,14 +1,20 @@
 package com.kimtaehoonki.task.milestone.application;
 
+import com.kimtaehoonki.task.exception.impl.AuthorizedException;
 import com.kimtaehoonki.task.exception.impl.MilestoneNotFoundException;
 import com.kimtaehoonki.task.exception.impl.ProjectNotFoundException;
 import com.kimtaehoonki.task.exception.impl.StartDateLaterThanEndDateException;
+import com.kimtaehoonki.task.member.AccountRestTemplate;
 import com.kimtaehoonki.task.milestone.domain.Milestone;
 import com.kimtaehoonki.task.milestone.domain.MilestoneRepository;
+import com.kimtaehoonki.task.milestone.dto.MilestoneResponseDto;
 import com.kimtaehoonki.task.milestone.presentation.dto.RegisterMilestoneRequestDto;
+import com.kimtaehoonki.task.project.domain.MemberInProjectRepository;
 import com.kimtaehoonki.task.project.domain.ProjectRepository;
 import com.kimtaehoonki.task.project.domain.entity.Project;
+import com.kimtaehoonki.task.project.presentation.dto.GetMilestonesByProjectIdResponseDto;
 import java.time.LocalDate;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,18 +25,29 @@ import org.springframework.transaction.annotation.Transactional;
 public class MilestoneServiceImpl implements MilestoneService {
     private final MilestoneRepository milestoneRepository;
     private final ProjectRepository projectRepository;
+    private final AccountRestTemplate accountRt;
+    private final MemberInProjectRepository memberInProjectRepository;
 
     @Override
     @Transactional
-    public Long registerMilestone(RegisterMilestoneRequestDto dto) {
+    public Long registerMilestone(RegisterMilestoneRequestDto dto, Integer memberId) {
         LocalDate startDate = dto.getStartDate();
         LocalDate endDate = dto.getEndDate();
 
         if (startDate.isAfter(endDate)) {
             throw new StartDateLaterThanEndDateException();
         }
+
+        accountRt.validateMemberExists(memberId);
+
         Project project = projectRepository.findById(dto.getProjectId())
             .orElseThrow(ProjectNotFoundException::new);
+
+        boolean isMemberInProject =
+            memberInProjectRepository.existsByProject_idAndMemberId(project.getId(), memberId);
+        if (!isMemberInProject) {
+            throw new AuthorizedException();
+        }
 
         Milestone milestone =
             Milestone.create(project, dto.getName(), startDate, endDate);
@@ -40,9 +57,28 @@ public class MilestoneServiceImpl implements MilestoneService {
 
     @Override
     @Transactional
-    public void deleteMilestone(Long milestoneId) {
+    public void deleteMilestone(Long milestoneId, Integer memberId) {
+        accountRt.validateMemberExists(memberId);
+
         Milestone milestone = milestoneRepository.findById(milestoneId)
             .orElseThrow(MilestoneNotFoundException::new);
+
+        Long projectId = milestone.getProject().getId();
+        Project project = projectRepository.findById(projectId)
+            .orElseThrow(ProjectNotFoundException::new);
+        boolean isMemberInProject =
+            memberInProjectRepository.existsByProject_idAndMemberId(project.getId(), memberId);
+        if (!isMemberInProject) {
+            throw new AuthorizedException();
+        }
+
         milestoneRepository.delete(milestone);
+    }
+
+    @Override
+    public GetMilestonesByProjectIdResponseDto getMilestonesByProjectId(Long projectId) {
+        List<MilestoneResponseDto> milestones = milestoneRepository.findAllByProject_id(projectId);
+        return new GetMilestonesByProjectIdResponseDto(milestones);
+
     }
 }
